+++ astropy/nddata/mixins/tests/test_ndarithmetic.py
@@ -10,10 +10,11 @@ from astropy.nddata.nduncertainty import (StdDevUncertainty, VarianceUncertainty
                               UnknownUncertainty,
                               IncompatibleUncertaintiesException)
 from astropy.nddata import NDDataRef
-from astropy.nddata.nddata import NDData
+from astropy.nddata import _testing as nd_testing
 
 from astropy.units import UnitsError, Quantity
 from astropy import units as u
+from astropy.wcs import WCS
 
 
 # Alias NDDataAllMixins in case this will be renamed ... :-)
@@ -184,10 +185,10 @@ def test_arithmetics_data_unit_not_identical(data1, data2):
 # None set
 @pytest.mark.parametrize(('wcs1', 'wcs2'), [
     (None, None),
-    (None, 5),
-    (5, None),
-    (5, 5),
-    (7, 5),
+    (None, WCS(naxis=2)),
+    (WCS(naxis=2), None),
+    nd_testing.create_two_equal_wcs(naxis=2),
+    nd_testing.create_two_unequal_wcs(naxis=2),
     ])
 def test_arithmetics_data_wcs(wcs1, wcs2):
 
@@ -205,16 +206,16 @@ def test_arithmetics_data_wcs(wcs1, wcs2):
 
     # Addition
     nd3 = nd1.add(nd2)
-    assert ref_wcs == nd3.wcs
+    nd_testing.assert_wcs_seem_equal(ref_wcs, nd3.wcs)
     # Subtraction
     nd4 = nd1.subtract(nd2)
-    assert ref_wcs == nd3.wcs
+    nd_testing.assert_wcs_seem_equal(ref_wcs, nd4.wcs)
     # Multiplication
     nd5 = nd1.multiply(nd2)
-    assert ref_wcs == nd3.wcs
+    nd_testing.assert_wcs_seem_equal(ref_wcs, nd5.wcs)
     # Division
     nd6 = nd1.divide(nd2)
-    assert ref_wcs == nd3.wcs
+    nd_testing.assert_wcs_seem_equal(ref_wcs, nd6.wcs)
     for nd in [nd3, nd4, nd5, nd6]:
         # Check all other attributes are not set
         assert nd.unit is None
@@ -1021,8 +1022,7 @@ def test_arithmetics_handle_switches(use_abbreviation):
     mask2 = False
     uncertainty1 = StdDevUncertainty([1, 2, 3])
     uncertainty2 = StdDevUncertainty([1, 2, 3])
-    wcs1 = 5
-    wcs2 = 100
+    wcs1, wcs2 = nd_testing.create_two_unequal_wcs(naxis=1)
     data1 = [1, 1, 1]
     data2 = [1, 1, 1]
 
@@ -1044,7 +1044,7 @@ def test_arithmetics_handle_switches(use_abbreviation):
     nd_ = nd3.add(nd2, propagate_uncertainties=False,
                   handle_meta=use_abbreviation, handle_mask=use_abbreviation,
                   compare_wcs=use_abbreviation)
-    assert nd_.wcs == wcs2
+    nd_testing.assert_wcs_seem_equal(nd_.wcs, wcs2)
     assert nd_.meta == meta2
     assert nd_.mask == mask2
     assert_array_equal(nd_.uncertainty.array, uncertainty2.array)
@@ -1053,7 +1053,7 @@ def test_arithmetics_handle_switches(use_abbreviation):
     nd_ = nd1.add(nd3, propagate_uncertainties=False,
                   handle_meta=use_abbreviation, handle_mask=use_abbreviation,
                   compare_wcs=use_abbreviation)
-    assert nd_.wcs == wcs1
+    nd_testing.assert_wcs_seem_equal(nd_.wcs, wcs1)
     assert nd_.meta == meta1
     assert nd_.mask == mask1
     assert_array_equal(nd_.uncertainty.array, uncertainty1.array)
@@ -1072,14 +1072,12 @@ def test_arithmetics_meta_func():
     mask2 = False
     uncertainty1 = StdDevUncertainty([1, 2, 3])
     uncertainty2 = StdDevUncertainty([1, 2, 3])
-    wcs1 = 5
-    wcs2 = 100
     data1 = [1, 1, 1]
     data2 = [1, 1, 1]
 
-    nd1 = NDDataArithmetic(data1, meta=meta1, mask=mask1, wcs=wcs1,
+    nd1 = NDDataArithmetic(data1, meta=meta1, mask=mask1,
                            uncertainty=uncertainty1)
-    nd2 = NDDataArithmetic(data2, meta=meta2, mask=mask2, wcs=wcs2,
+    nd2 = NDDataArithmetic(data2, meta=meta2, mask=mask2,
                            uncertainty=uncertainty2)
 
     nd3 = nd1.add(nd2, handle_meta=meta_fun_func)
@@ -1096,10 +1094,9 @@ def test_arithmetics_meta_func():
 
 def test_arithmetics_wcs_func():
     def wcs_comp_func(wcs1, wcs2, tolerance=0.1):
-        if abs(wcs1 - wcs2) <= tolerance:
-            return True
-        else:
+        if tolerance < 0.01:
             return False
+        return True
 
     meta1 = {'a': 1}
     meta2 = {'a': 3, 'b': 2}
@@ -1107,8 +1104,7 @@ def test_arithmetics_wcs_func():
     mask2 = False
     uncertainty1 = StdDevUncertainty([1, 2, 3])
     uncertainty2 = StdDevUncertainty([1, 2, 3])
-    wcs1 = 99.99
-    wcs2 = 100
+    wcs1, wcs2 = nd_testing.create_two_equal_wcs(naxis=1)
     data1 = [1, 1, 1]
     data2 = [1, 1, 1]
 
@@ -1118,11 +1114,14 @@ def test_arithmetics_wcs_func():
                            uncertainty=uncertainty2)
 
     nd3 = nd1.add(nd2, compare_wcs=wcs_comp_func)
-    assert nd3.wcs == 99.99
+    nd_testing.assert_wcs_seem_equal(nd3.wcs, wcs1)
 
+    # Fails because the function fails
     with pytest.raises(ValueError):
         nd1.add(nd2, compare_wcs=wcs_comp_func, wcs_tolerance=0.00001)
 
+    # Fails because for a parameter to be passed correctly to the function it
+    # needs the wcs_ prefix
     with pytest.raises(KeyError):
         nd1.add(nd2, compare_wcs=wcs_comp_func, tolerance=1)
 
@@ -1140,14 +1139,12 @@ def test_arithmetics_mask_func():
     mask2 = [True, False, False]
     uncertainty1 = StdDevUncertainty([1, 2, 3])
     uncertainty2 = StdDevUncertainty([1, 2, 3])
-    wcs1 = 99.99
-    wcs2 = 100
     data1 = [1, 1, 1]
     data2 = [1, 1, 1]
 
-    nd1 = NDDataArithmetic(data1, meta=meta1, mask=mask1, wcs=wcs1,
+    nd1 = NDDataArithmetic(data1, meta=meta1, mask=mask1,
                            uncertainty=uncertainty1)
-    nd2 = NDDataArithmetic(data2, meta=meta2, mask=mask2, wcs=wcs2,
+    nd2 = NDDataArithmetic(data2, meta=meta2, mask=mask2,
                            uncertainty=uncertainty2)
 
     nd3 = nd1.add(nd2, handle_mask=mask_sad_func)